+++ Source/ACE/Entity/Monster.cs
@@ -13,7 +13,7 @@ namespace ACE.Entity
           : base(aceO)
         {
             // TODO: Check why Drudges don't appear on radar yet and don't have a healthbar when you select them
-            aceO.AceObjectId = new ObjectGuid(GuidManager.NewMonsterGuid()).Full;
+            aceO.AceObjectId = new ObjectGuid(GuidManager.NewNonStaticGuid()).Full;
             IsAlive = true;
             SetupVitals();
         }