+++ Source/ACE/Entity/Player.cs
@@ -21,6 +21,10 @@ using System.IO;
 
 namespace ACE.Entity
 {
+    using System.Diagnostics;
+
+    using global::ACE.Factories;
+
     public sealed class Player : Creature
     {
         private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
@@ -382,7 +386,7 @@ namespace ACE.Entity
                 bool trainNewSkill = Character.TrainSkill(skill, creditsSpent);
 
                 // create an update to send to the client
-                var currentCredits = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.AvailableSkillCredits, Character.AvailableSkillCredits);
+                var currentCredits = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.AvailableSkillCredits, Character.AvailableSkillCredits);
 
                 // as long as the skill is sent, the train new triangle button on the client will not lock up.
                 // Sending Skill.None with status untrained worked in test
@@ -454,12 +458,12 @@ namespace ACE.Entity
                 string level = $"{Character.Level}";
                 string skillCredits = $"{Character.AvailableSkillCredits}";
                 string xpAvailable = $"{Character.AvailableExperience:#,###0}";
-                var levelUp = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.Level, Character.Level);
+                var levelUp = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.Level, Character.Level);
                 string levelUpMessageText = (Character.Level == maxLevel.Level) ? $"You have reached the maximum level of {level}!" : $"You are now level {level}!";
                 var levelUpMessage = new GameMessageSystemChat(levelUpMessageText, ChatMessageType.Advancement);
                 string xpUpdateText = (Character.AvailableSkillCredits > 0) ? $"You have {xpAvailable} experience points and {skillCredits} skill credits available to raise skills and attributes." : $"You have {xpAvailable} experience points available to raise skills and attributes.";
                 var xpUpdateMessage = new GameMessageSystemChat(xpUpdateText, ChatMessageType.Advancement);
-                var currentCredits = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.AvailableSkillCredits, Character.AvailableSkillCredits);
+                var currentCredits = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.AvailableSkillCredits, Character.AvailableSkillCredits);
                 if (Character.Level != maxLevel.Level && !creditEarned)
                 {
                     string nextCreditAtText = $"You will earn another skill credit at {chart.Levels.Where(item => item.Level > Character.Level).OrderBy(item => item.Level).First(item => item.GrantsSkillPoint).Level}";
@@ -775,9 +779,9 @@ namespace ACE.Entity
             // create and send the client death event, GameEventYourDeath
             var msgYourDeath = new GameEventYourDeath(Session, $"You have {currentDeathMessage}");
             var msgHealthUpdate = new GameMessagePrivateUpdateAttribute2ndLevel(Session, Vital.Health, Health.Current);
-            var msgNumDeaths = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.NumDeaths, Character.NumDeaths);
-            var msgDeathLevel = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.DeathLevel, Character.DeathLevel);
-            var msgVitaeCpPool = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.VitaeCpPool, Character.VitaeCpPool);
+            var msgNumDeaths = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.NumDeaths, Character.NumDeaths);
+            var msgDeathLevel = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.DeathLevel, Character.DeathLevel);
+            var msgVitaeCpPool = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.VitaeCpPool, Character.VitaeCpPool);
             var msgPurgeEnchantments = new GameEventPurgeAllEnchantments(Session);
             // var msgDeathSound = new GameMessageSound(Guid, Sound.Death1, 1.0f);
 
@@ -1137,7 +1141,7 @@ namespace ACE.Entity
         }
 
         /// <summary>
-        /// Set the currenly position of the character, to later save in the database.
+        /// Set the currently position of the character, to later save in the database.
         /// </summary>
         public void SetPhysicalCharacterPosition()
         {
@@ -1203,7 +1207,7 @@ namespace ACE.Entity
         {
             try
             {
-                Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.Age, Character.Age));
+                Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.Age, Character.Age));
             }
             catch (NullReferenceException)
             {
@@ -1348,7 +1352,7 @@ namespace ACE.Entity
         {
             Session.Player.Burden = UpdateBurden();
             Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(
-                                            Session,
+                                            Session.Player.Sequences,
                                             PropertyInt.EncumbranceVal,
                                             Session.Player.Burden ?? 0u));
         }
@@ -1516,7 +1520,7 @@ namespace ACE.Entity
                 // FIXME(ddevec): I should probably make a better interface for this
                 UpdateVitalInternal(Mana, Mana.Current / 2);
 
-                var updateCombatMode = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.CombatMode, 1);
+                var updateCombatMode = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.CombatMode, 1);
 
                 var motionLifestoneRecall = new UniversalMotion(MotionStance.Standing, new MotionItem(MotionCommand.LifestoneRecall));
 
@@ -1563,7 +1567,7 @@ namespace ACE.Entity
 
             var sysChatMessage = new GameMessageSystemChat(message, ChatMessageType.Recall);
 
-            var updateCombatMode = new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.CombatMode, 1);
+            var updateCombatMode = new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.CombatMode, 1);
 
             var motionMarketplaceRecall = new UniversalMotion(MotionStance.Standing, new MotionItem(MotionCommand.MarketplaceRecall));
 
@@ -1627,7 +1631,7 @@ namespace ACE.Entity
             inContainerChain.AddAction(this, () => Session.Network.EnqueueSend(new GameMessageCreateObject(item),
                 new GameMessageUpdateInstanceId(item.Guid, container.Guid, PropertyInstanceId.Container),
                 new GameMessageUpdateInstanceId(container.Guid, new ObjectGuid(0), PropertyInstanceId.Wielder),
-                new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.CurrentWieldedLocation, 0)));
+                new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.CurrentWieldedLocation, 0)));
 
             CurrentLandblock.EnqueueBroadcast(Location, Landblock.MaxObjectRange,
                                                 new GameMessageObjDescEvent(this),
@@ -1658,6 +1662,88 @@ namespace ACE.Entity
             }
             Session.Network.EnqueueSend(new GameMessagePutObjectInContainer(Session, container.Guid, item, location),
                                         new GameMessageUpdateInstanceId(item.Guid, container.Guid, PropertyInstanceId.Container));
+        }
+
+        /// <summary>
+        /// This method is used to split a stack of any item that is stackable - arrows, tapers, pyreal etc.
+        /// </summary>
+        /// <param name="stackId"></param>
+        /// <param name="containerId"></param>
+        /// <param name="place"></param>
+        /// <param name="amount"></param>
+        /// <returns></returns>
+        public void HandleActionStackableSplitToContainer(uint stackId, uint containerId, uint place, ushort amount)
+        {
+            ActionChain splitItemsChain = new ActionChain();
+            splitItemsChain.AddAction(this, () =>
+                {
+                    Container container;
+                    if (containerId == Guid.Full)
+                    {
+                        container = this;
+                    }
+                    else
+                    {
+                        container = (Container)GetInventoryItem(new ObjectGuid(containerId));
+                    }
+
+                    if (container == null)
+                    {
+                        log.InfoFormat("Asked to split stack {0} in container {1} - the container was not found",
+                            stackId,
+                            containerId);
+                        return;
+                    }
+                    WorldObject stack = container.GetInventoryItem(new ObjectGuid(stackId));
+                    if (stack == null)
+                    {
+                        log.InfoFormat("Asked to split stack {0} in container {1} - the stack item was not found",
+                            stackId,
+                            containerId);
+                        return;
+                    }
+                    if (stack.Value == null || stack.StackSize < amount || stack.StackSize == 0)
+                    {
+                        log.InfoFormat(
+                            "Asked to split stack {0} in container {1} - with amount of {2} but there is not enough to split",
+                            stackId, containerId, amount);
+                        return;
+                    }
+
+                    // Ok we are in business
+
+                    WorldObject newStack = new DebugObject(stack.NewAceObjectFromCopy());
+                    container.AddToInventory(newStack);
+                    var valuePerItem = stack.Value / stack.StackSize;
+                    var burdenPerItem = stack.Burden / stack.StackSize;
+                    ushort oldStackSize = (ushort)stack.StackSize;
+                    ushort newStackSize = (ushort)stack.StackSize;
+                    newStackSize = (ushort)(newStackSize - amount);
+
+                    newStack.StackSize = amount;
+                    newStack.Value = newStack.StackSize * valuePerItem;
+                    newStack.Burden = (ushort)(newStack.StackSize * burdenPerItem);
+
+                    stack.StackSize = newStackSize;
+                    stack.Value = stack.StackSize * valuePerItem;
+                    stack.Burden = (ushort)(stack.StackSize * burdenPerItem);
+
+                    GameMessagePrivateUpdatePropertyInt msgUpdateValue =
+                        new GameMessagePrivateUpdatePropertyInt(container.Sequences, PropertyInt.Value, 1);
+                    GameMessagePutObjectInContainer msgPutObjectInContainer =
+                        new GameMessagePutObjectInContainer(Session, container.Guid, newStack, place);
+                    Debug.Assert(stack.StackSize != null, "stack.StackSize != null");
+                    Debug.Assert(stack.Value != null, "stack.Value != null");
+                    GameMessageSetStackSize msgAdjustOldStackSize = new GameMessageSetStackSize(stack.Sequences,
+                        stack.Guid, (int)stack.StackSize, oldStackSize);
+
+                    // TODO: Finish here - create the object, and finish up the code.   Wrap in ActionChain. Og II
+
+                    GameMessageCreateObject msgNewStack = new GameMessageCreateObject(newStack);
+                    CurrentLandblock.EnqueueBroadcast(Location, MaxObjectTrackingRange, msgUpdateValue,
+                        msgPutObjectInContainer, msgAdjustOldStackSize, msgNewStack);
+                });
+            splitItemsChain.EnqueueChain();
         }
 
         private void HandlePickupItem(Container container, ObjectGuid itemGuid, uint location, PropertyInstanceId iidPropertyId)
@@ -1712,7 +1798,7 @@ namespace ACE.Entity
                     if (iidPropertyId == PropertyInstanceId.Container)
                     {
                         Session.Network.EnqueueSend(
-                            new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.EncumbranceVal, UpdateBurden()),
+                            new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.EncumbranceVal, UpdateBurden()),
                             new GameMessageSound(Guid, Sound.PickUpItem, 1.0f),
                             new GameMessageUpdateInstanceId(itemGuid, container.Guid, iidPropertyId),
                             new GameMessagePutObjectInContainer(Session, container.Guid, item, location));
@@ -1879,7 +1965,7 @@ namespace ACE.Entity
                                         PropertyInstanceId.Container),
                                     new GameMessageUpdateInstanceId(container.Guid, itemGuid,
                                         PropertyInstanceId.Wielder),
-                                    new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.CurrentWieldedLocation,
+                                    new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.CurrentWieldedLocation,
                                         location));
                                 CurrentLandblock.EnqueueBroadcast(
                                     Location,
@@ -1896,7 +1982,7 @@ namespace ACE.Entity
                                     new GameMessageSound(Guid, Sound.WieldObject, (float)1.0),
                                     new GameMessageUpdateInstanceId(container.Guid, new ObjectGuid(0), PropertyInstanceId.Container),
                                     new GameMessageUpdateInstanceId(container.Guid, itemGuid, PropertyInstanceId.Wielder),
-                                    new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.CurrentWieldedLocation, location));
+                                    new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.CurrentWieldedLocation, location));
                                 CurrentLandblock.EnqueueBroadcast(
                                     Location,
                                     Landblock.MaxObjectRange,
@@ -1977,7 +2063,7 @@ namespace ACE.Entity
                 }
                 SetInventoryForWorld(inventoryItem);
                 RemoveFromInventory(itemGuid);
-                Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(Session, PropertyInt.EncumbranceVal, (uint)Burden));
+                Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(Session.Player.Sequences, PropertyInt.EncumbranceVal, (uint)Burden));
 
                 var motion = new UniversalMotion(MotionStance.Standing);
                 motion.MovementData.ForwardCommand = (ushort)MotionCommand.Pickup;
@@ -2417,7 +2503,7 @@ namespace ACE.Entity
 
             foreach (AceObjectPropertiesSpell x in propertiesSpellId)
             {
-                    writer.Write(x.SpellId);
+                writer.Write(x.SpellId);
             }
         }