+++ Source/ACE/Network/GameMessages/Messages/GameMessagePrivateUpdatePropertyInt.cs
@@ -2,12 +2,14 @@
 
 namespace ACE.Network.GameMessages.Messages
 {
+    using global::ACE.Network.Sequence;
+
     public class GameMessagePrivateUpdatePropertyInt : GameMessage
     {
-        public GameMessagePrivateUpdatePropertyInt(Session session, PropertyInt property, uint value)
+        public GameMessagePrivateUpdatePropertyInt(SequenceManager sequences, PropertyInt property, uint value)
             : base(GameMessageOpcode.PrivateUpdatePropertyInt, GameMessageGroup.Group09)
         {
-            Writer.Write(session.Player.Sequences.GetNextSequence(Sequence.SequenceType.PrivateUpdatePropertyInt));
+            Writer.Write(sequences.GetNextSequence(SequenceType.PrivateUpdatePropertyInt));
             Writer.Write((uint)property);
             Writer.Write(value);
         }